// src/shared/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  logistics: {
    Tables: {
// ...
      uom: {
        Row: {
          base_uom_code: string | null
          category: string
          conversion_factor: number | null
          created_at: string
          decimals: number | null
          id: string
          status: string
          tenant_id: string
          uom_code: string
          uom_name: string
          updated_at: string
        }
        Insert: {
          base_uom_code?: string | null
          category: string
          conversion_factor?: number | null
          created_at?: string
          decimals?: number | null
          id?: string
          status?: string
          tenant_id: string
          uom_code: string
          uom_name: string
          updated_at?: string
        }
        Update: {
          base_uom_code?: string | null
          category?: string
          conversion_factor?: number | null
          created_at?: string
          decimals?: number | null
          id?: string
          status?: string
          tenant_id?: string
          uom_code?: string
          uom_name?: string
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}